Updated test
